requirements.txt
python-telegram-bot==5.3.0
praw==5.3.0